http://malideveloper.arm.com/resources/sdks/mali-opencl-sdk/ Software Development Kit OpenCL™ on ARM Linux
The Mali OpenCL SDK provides developers with a framework and series of samples for developing OpenCL 1.1 applications on ARM Mali based platforms such as the Mali-T600 and above family of GPUs. The samples cover a wide range of use cases that utilize the Mali GPU to achieve a significant improvement in performance when compared to running on the CPU alone.